feat: Use electron integration, Refactor code
